Refactored code to make more sense in atomic structure
