+++ .gitignore
@@ -9,4 +9,4 @@ node_modules/
 /coverage
 
 # production
-client/build/**
+client/build/
