+++ client/src/components/UI/molecule/card.jsx
@@ -1,6 +1,6 @@
 import { Link } from "react-router-dom";
 
-import { toHeading } from "../atom/toHeading";
+import { toHeading } from "../../utils/toHeading";
 
 import styles from "./styles/card.module.css";
