+++ client/src/components/UI/organism/Header.jsx
@@ -1,5 +1,5 @@
-import { toLink } from "../atom/toLink";
-import { toHeading } from "../atom/toHeading";
+import { toLink } from "../../utils/toLink";
+import { toHeading } from "../../utils/toHeading";
 
 import styles from "./styles/Header.module.css";
 import { useAuth0 } from "@auth0/auth0-react";
