+++ client/src/pages/styles/Home.module.css
@@ -37,6 +37,7 @@
   .home_content {
     padding: 50px 80px;
   }
+
   .showcase .container {
     margin-top: 15vh;
   }
